==> notebooks-scripts/01-creando-y-generando-archivos-de-entrada-in.ipynb <==
# Creando y generando archivos de entrada .in para Quantum Expresso

## Importando liberías a utilizar

import os
import numpy as np

## Definiendo volúmenes a utilizar $V$  $(\mathring{A}^3)$

volumenes = [159.30, 162.84, 166.38, 169.92, 173.46, 177.00, 180.54, 184.08, 187.62, 191.16, 194.70]

## Definiendo cocientes entre lattice parámetros $(c/a)$

c_a_coeficientes = [0.986, 1.008, 1.030, 1.052, 1.074, 1.096, 1.118, 1.140, 1.162, 1.183, 1.206]

## Definiendo función para obtener el valor del lattice parametro $a$ a partir de la siguiente ecuación

$$
V = \frac{\sqrt{3}}{2}a^3(c/a)\quad \Rightarrow \quad a= \sqrt[3]{\frac{2V}{\sqrt{3}\cdot(c/a)}}
$$

Considerando que conocemos el valor de $V$ y $c/a$.

def a(V:float,c_a:float) -> float:
    return round(np.cbrt((2 * V) / (np.sqrt(3) * c_a)),5)

### Obteniendo valores de $a$ por cada volumen y por cada valor de $c/a$ y añadiendolos a una nueva lista con entradas $(V,c/a,a)$.

V_c_a_a_s = list()

for volumen in volumenes:
    for c_a_coeficiente in c_a_coeficientes:
        V_c_a_a_s.append([volumen,c_a_coeficiente,a(volumen,c_a_coeficiente)])

V_c_a_a_s

## Imprimiendo nuestro texto modelo para crear archivos .in

### Cambiando de directorio donde se encuentra nuestro archivo mob2_pw.in, el cual nos servira como modelo para la creación de todos los demás.

raw_path = '../data/raw'
raw_dir = os.chdir(raw_path)
raw_files = os.listdir(raw_dir)
raw_files

### Imprimiendo en pantalla el texto que contiene.

pw_file = open(raw_files[0],'r')
pw_read = pw_file.read()
print(pw_read)

# Cerrando archivo anterior
pw_file.close()

### Regresando al directorio actual

actual_dir = '../../notebooks'
os.chdir(actual_dir)

## Creando sistema de directorios en interim y archivos pw.in

### Diseño en forma de árbol

A manera de ejemplo, el sistema de archivos quedaría dado de la siguiente manera para cada uno de los volumenes.

```bash
../data/
├── .interim
    └── .159_30
        └── .0_986
            └── .temp
        └── .1_008
        └── .1_030
        └── .1_052
        └── .1_074
        └── .1_096
        └── .1_118
        └── .1_140
        └── .1_162
        └── .1_183
        └── .1_206
```

## Creando archivos archivo pw.in

def creating_mob2_file(V_c_a_a_1,V_c_a_a_2):
    crystal = r'{crystal}'
    mob2_pw_file  = f''' &CONTROL
    prefix='mob2',
    pseudo_dir='/home/marcocarmonaga/Apps/qe-7.0/pseudo',
    outdir='./tmp',
 /
 &SYSTEM
    ibrav =  4,
    celldm(1) = {V_c_a_a_2}
    celldm(3) = {V_c_a_a_1}
    nat =  3,
    ntyp = 2,
    nbnd = 20,
    ecutwfc = 36.0,
    occupations = 'smearing',
    smearing = 'gauss',
    degauss = 0.015,
 /
 &ELECTRONS
 /
ATOMIC_SPECIES
   Mo  95.95  Mo_ONCV_PBE-1.2.upf
   B   10.81  B_ONCV_PBE-1.2.upf
ATOMIC_POSITIONS {crystal}
   Mo 0.00 0.00 0.00
   B 0.666666667  0.333333333  0.5
   B 0.333333333  0.666666667  0.5
K_POINTS automatic
   18 18 18   0 0 1
'''
    return mob2_pw_file

mob2_pw_files = list()

for V_c_a_a in V_c_a_a_s:
   mob2_pw_files.append(creating_mob2_file(V_c_a_a_1=V_c_a_a[1],V_c_a_a_2=V_c_a_a[2]))

### Uniendo valores $(V,c/a,a,\text{mob2-pw-file})$

for V_c_a_a in V_c_a_a_s:
    V_c_a_a.append(mob2_pw_files[V_c_a_a_s.index(V_c_a_a)])

V_c_a_a_mob2_file_s = V_c_a_a_s.copy()

### Creando directorios completos

os.chdir('../data/interim')
for V_c_a_a_mob2_file in V_c_a_a_mob2_file_s:
    directorio_primario = str(V_c_a_a_mob2_file[0]).replace('.','_')
    directorio_secundario = str(V_c_a_a_mob2_file[1]).replace('.','_')
    try:
        os.mkdir(directorio_primario)
        os.chdir(directorio_primario)
        os.mkdir(directorio_secundario)
        os.chdir(directorio_secundario)
        os.mkdir('tmp')
        with open('mob2_pw.in','w') as pw_file:
            pw_file.writelines(V_c_a_a_mob2_file[3])
            pw_file.close()
        os.chdir('../../')
    except:
        os.chdir(directorio_primario)
        os.mkdir(directorio_secundario)
        os.chdir(directorio_secundario)
        os.mkdir('tmp')
        with open('mob2_pw.in','w') as pw_file:
            pw_file.writelines(V_c_a_a_mob2_file[3])
            pw_file.close()
        os.chdir('../../')
os.chdir('../../notebooks')
