==> notebooks-scripts/03-obteniendo-datos-de-energia-total.ipynb <==
# Obteniendo datos de la energia total por cada volumen

## Importando liberías a utilizar

import os
import pandas as pd

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

energia_total_directory = '../data/interim'
os.chdir(energia_total_directory)

## Extrayendo linea de energia total y guardandola en una lista.

directorios_volumenes = os.listdir()

energia_total_por_volumenes = list()

for directorio_volumen in directorios_volumenes:
    energia_total_por_volumen = list()
    os.chdir(directorio_volumen)
    sub_directorios = os.listdir()
    for sub_directorio in sub_directorios:
        energia_total = list()
        os.chdir(sub_directorio)
        with open('mob2_pw.out','r') as file:
            for line in file:
                if '!' in line:
                    line = line[36:49]
                    energia_total.append(line)
        energia_total_por_volumen.append(energia_total)
        os.chdir('..')
    energia_total_por_volumenes.append(energia_total_por_volumen)
    os.chdir('..')

energia_total_por_volumenes

## Uniendo todo en una sola lista $(V,c/a,E)$

V_c_a_E_s = list()

c_a_s = [float(c_a.replace('_','.')) for c_a in sub_directorios]

for V,c_a,E in zip(directorios_volumenes,sub_directorios,energia_total_por_volumenes):
    V = float(V.replace('_','.'))
    c_a = float(c_a.replace('_','.'))
    E = [float(i) for e in E for i in e]
    V_c_a_E_s.append([V,c_a_s,E])

V_c_a_E_s

## Creando archivos csv

os.chdir('../processed/energia-total-por-volumen')

for V_c_a_E in V_c_a_E_s:
    df = pd.DataFrame()
    df['c_a_value'] = V_c_a_E[1]
    df['Energy'] = V_c_a_E[2]
    df = df.sort_values(by='c_a_value')
    name_file = str(V_c_a_E[0]).replace('.','_')
    df.to_csv(f'{name_file}.csv',index=False)

## Regresando a directorio principal

os.chdir(directorio_principal)
