==> notebooks-scripts/08-obteniendo-valor-de-c_a-para-v0.ipynb <==
# Obteniendo valor de $c/a$ y Energía, para el valor de $V0=177.91$

## Importando liberías a utilizar

import os
import pandas as pd
import numpy as np

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

energia_total_directory = '../data/interim'
os.chdir(energia_total_directory)

## Listando archivos

files = os.listdir()
files

## Creando carpeta para $V=177.91$

try:
    os.mkdir('177_91')
    os.chdir('177_91')
except:
    os.chdir('177_91')

## Creando subcarpetas

subdirectorios = ['0_986','1_03','1_008','1_14','1_052','1_074','1_096','1_118','1_162','1_183','1_206']

try:
    for subdirectorio in subdirectorios:
        os.mkdir(subdirectorio)
        os.chdir(subdirectorio)
        os.mkdir('tmp')
        os.chdir('..')
except:
    pass

## Definiendo función para obtener el valor del lattice parametro $a$ a partir de la siguiente ecuación

$$
V = \frac{\sqrt{3}}{2}a^3(c/a)\quad \Rightarrow \quad a= \sqrt[3]{\frac{2V}{\sqrt{3}\cdot(c/a)}}
$$

Considerando que conocemos el valor de $V$ y $c/a$.

def a(V:float,c_a:float) -> float:
    return round(np.cbrt((2 * V) / (np.sqrt(3) * c_a)),5)

### Encontrando valores de a para cada valor de $c/a$ siguiente

c_a_s = [0.986,1.03,1.008,1.14,1.052,1.074,1.096,1.118,1.162,1.183,1.206]

c_a_a_s = list()

for c_a in c_a_s:
    a_value = a(177.91,c_a)
    c_a_a_s.append([c_a,a_value])

c_a_a_s

## Creando archivos archivo pw.in

def creating_mob2_file(c_a_a):
    crystal = r'{crystal}'
    mob2_pw_file  = f''' &CONTROL
    prefix='mob2',
    pseudo_dir='/home/marcocarmonaga/Apps/qe-7.0/pseudo',
    outdir='./tmp',
 /
 &SYSTEM
    ibrav =  4,
    celldm(1) = {c_a_a[1]}
    celldm(3) = {c_a_a[0]}
    nat =  3,
    ntyp = 2,
    nbnd = 20,
    ecutwfc = 36.0,
    occupations = 'smearing',
    smearing = 'gauss',
    degauss = 0.015,
 /
 &ELECTRONS
 /
ATOMIC_SPECIES
   Mo  95.95  Mo_ONCV_PBE-1.2.upf
   B   10.81  B_ONCV_PBE-1.2.upf
ATOMIC_POSITIONS {crystal}
   Mo 0.00 0.00 0.00
   B 0.666666667  0.333333333  0.5
   B 0.333333333  0.666666667  0.5
K_POINTS automatic
   18 18 18   0 0 1
'''
    return mob2_pw_file

c_a_s_pw_files = list()

for c_a_a in c_a_a_s:
   c_a_s_pw_files.append([c_a_a[0],c_a_a[1], creating_mob2_file(c_a_a)])

c_a_s_pw_files

## Guardando archivos en respectivas carpetas

os.getcwd()

for c_a_s_pw_file in c_a_s_pw_files:
    subdirectory_name = str(c_a_s_pw_file[0]).replace('.','_')
    os.chdir(subdirectory_name)
    with open('mob2_pw.in','w') as file:
        file.writelines(c_a_s_pw_file[2])
        file.close()
    os.chdir('..')

## Regresando a directorio principal

os.chdir(directorio_principal)
