==> notebooks-scripts/12-obteniendo-valores-minimos-para-v_177_91.ipynb <==
# Obteniendo valores minimos para $c/a$ con su respectivo valor de energía.

## Importando liberías a utilizar

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

energia_total_directory = '../data/processed/energia-total-por-volumen'
os.chdir(energia_total_directory)

## Enlistando archivos dentro de la carpeta secundaria

files = os.listdir()
files

## Ajustando puntos de Energía Total a polinomio de 4to grado, derivando y obteniendo valores mínimos

valores_minimos = list()

for file in files:
    if file == '177_91.csv':
        df = pd.read_csv(file)
        c_a_valores = df['c_a_value'].to_list()
        c_a = np.array(c_a_valores)
        energia = df['Energy'].to_list()
        energia_total = np.array(energia)
        ajuste = np.polyfit(c_a,energia_total,deg=4)
        p = np.poly1d(ajuste)
        p_deriv = np.polyder(p)
        roots = np.roots(p_deriv)
        root = roots[0].real
        c_a_min = root.round(5)
        energia_min = p(root).round(5)
        volumen_str = file.rpartition('.')
        volumen = float(volumen_str[0].replace('_','.'))
        valores_minimos.append([volumen,c_a_min,energia_min])

valores_minimos_df = pd.DataFrame(valores_minimos, columns=['Volumen','c_a_min','Energia_min'])
valores_minimos_df

## Concatenando valor a archivo `valores_minimos.csv`

df_2 = pd.read_csv('../valores_minimos.csv')
main_df = pd.concat([df_2,valores_minimos_df])
main_df.sort_values(by='Volumen',inplace=True)

## Exportando main_df a csv

main_df.to_csv('../valores-minimos-uniendo-177_09.csv',index=False)

## Regresando a directorio principal

os.chdir(directorio_principal)
