==> notebooks-scripts/13-obteniendo-parametros-alpha-y-betha.ipynb <==
# Encontrando parámetros $\alpha$ y $\beta$ que se ajusten a la función

$$
(c/a)(V) = (c/a)(V_0) + \alpha(1-V/V_0) + \beta(1-V/V_0)^2,
$$

de los datos en la tabla `data/processed/vol_energia_presion.csv`.

## Importando liberías a utilizar

import os
import pandas as pd
import numpy as np

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

energia_total_directory = '../data/processed/'
os.chdir(energia_total_directory)

## Listando archivos

files = os.listdir()
files

## Definiendo variables a utilizar

# V_0
V0 = 177.91

# (c/a)(V0)
c_a_V0 = 1.10154

## Encontrando parámetros $\alpha$ y $\beta$

La idea está en encontrar los parámetros $\alpha$, $\beta$; tales que se cumpla lo siguiente:

$$
\begin{pmatrix}
1-V_{i}/V_0 & (1-V_{i}/V_0)^2\\
\vdots & \vdots
\end{pmatrix}
\begin{pmatrix}
\alpha \\
\beta
\end{pmatrix}=
\begin{pmatrix}
(c/a)(V_{i}) - (c/a)(V_0)\\
\vdots
\end{pmatrix}
$$

### Obteniendo cada uno de los términos

df = pd.read_csv('valores-minimos-uniendo-177_09.csv')
df.drop(columns=['Energia_min'], inplace=True)
df['termino_lineal'] = df['Volumen'].apply(lambda x: 1 - (x / V0))
df['termino_cuadratico'] = df['Volumen'].apply(lambda x: (1 - (x / V0))**2)
df['diferencia'] = df['c_a_min'].apply(lambda x: x-c_a_V0)
df

## Eliminando columnas innecesarias para el ajuste.

df_copy = df.copy()
df_copy.drop(columns=['Volumen','c_a_min'],inplace=True)

## Separando dataframe y convirtiendo en matriz

A = df_copy.iloc[:,0:2].values
b = df_copy['diferencia'].values

print(A,b)

## Entontrar parámetros $\alpha$ y $\beta$ resolviendo lo siguiente:

$$
x = A^{-1}b
$$

donde $x=[\alpha\ \beta]$

Sin embargo, al tratarse de una matriz no cuadrada, procederemos a hacer una descomposición QR, para así proceder a resolver el siguiente sistema:

$$
x = R^{-1}Q^{T}b
$$

Q, R = np.linalg.qr(A)

alpha, beta = np.linalg.solve(R, np.dot(Q.T,b))
print('alpha =',alpha.round(5))
print('beta =',beta.round(5))

## Conclusión

$$
\boxed{\therefore\ (c/a)(V) = (c/a)(V_0) + \alpha(1-V/V_0) + \beta(1-V/V_0)^2}
$$

con $\alpha = -0.18024$, $\beta = 0.13883$ y $(c/a)(V_0)=1.10154$

## Regresando a directorio principal

os.chdir(directorio_principal)
