==> notebooks-scripts/15-construyendo-tabla-presion-volumen-valores-especificos.ipynb <==
# Construyendo tabla con columnas: Presion, Volumen, c_a
## Para valores específicos

donde los valores de presión van desde los **0 GPa** hasta **150 GPa**.

considerando las siguientes ecuaciones:

$$
P(V) = \frac{3B_0}{2}\left[ \left(\frac{V_{0}}{V}\right)^{\frac{7}{3}} -\left(\frac{V_{0}}{V}\right)^{\frac{5}{3}} \right]\left\{1+\frac{3}{4}\left(B_0^\prime-4\right)\left[ \left(\frac{V_{0}}{V}\right)^{\frac{7}{3}} -1 \right]\right\}
$$

y

$$
(c/a)(V) = (c/a)(V_0) + \alpha(1-V/V_0) + \beta(1-V/V_0)^2
$$

con $\alpha = -0.18024$, $\beta = 0.13883$ y $(c/a)(V_0)=1.10154$


## Importando liberías a utilizar

import os
import pandas as pd
import numpy as np
from scipy import optimize

## Definiendo variables a utilizar

V0 = 177.91
B0 = 299.8
B0_prima = 4.12
c_a_V0 = 1.10154
alpha = -0.18024
beta = 0.13883

## Definiendo funciones a utilizar

def P(V):
    return (3/2) * B0 * (((V0 / V)**(7/3)) - ((V0 / V)**(5/3))) * (1 + (3/4) * (B0_prima - 4) * (((V0 / V)**(7/3)) - 1))

def c_a(V):
    return c_a_V0 + alpha*(1 - (V / V0)) + beta*((1 - (V / V0))**2)

## Construyendo dataframe

df = pd.DataFrame()
df['Presion_GPa'] = [0,50,65,70,80,90,100,110,120,130,140,150]
df

presion_valores = df['Presion_GPa'].to_list()

volumenes = list()

for presion_valor in presion_valores:
    def Presion(V):
        return (3/2) * B0 * (((V0 / V)**(7/3)) - ((V0 / V)**(5/3))) * (1 + (3/4) * (B0_prima - 4) * (((V0 / V)**(7/3)) - 1)) - presion_valor
    
    volumenes.append(round(optimize.fsolve(Presion,1)[0],5))

df['Volumenes'] = volumenes
df['c_a'] = df['Volumenes'].apply(lambda x: round(c_a(x),5))

df

## Exportando csv

df.to_csv('../data/processed/presion-Volumen-c_a-valores-especificos.csv',index=False)
