==> notebooks-scripts/16-obteniendo-valores-de-presion.ipynb <==
# Obteniendo valores de presión para cada volumen, evaluando la ecuación de estado de Birch–Murnaghan

$$
P(V) = \frac{3B_0}{2}\left[ \left(\frac{V_{0}}{V}\right)^{\frac{7}{3}} -\left(\frac{V_{0}}{V}\right)^{\frac{5}{3}} \right]\left\{1+\frac{3}{4}\left(B_0^\prime-4\right)\left[ \left(\frac{V_{0}}{V}\right)^{\frac{7}{3}} -1 \right]\right\}
$$

## Importando liberías a utilizar

import os
import pandas as pd
import numpy as np

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

energia_total_directory = '../data/processed/'
os.chdir(energia_total_directory)

## Definiendo variables ($V_0$, $B_0$, $B_0^\prime$) a partir del archivo `data/processed/eos/eos_out.txt`

V0 = 177.91
B0 = 299.8
B0_prima = 4.12
c_a_V0 = 1.10154
alpha = -0.18024
beta = 0.13883

## Definiendo ecuación de estado de Birch–Murnaghan

def P(V):
    return round((3/2) * B0 * (((V0 / V)**(7/3)) - ((V0 / V)**(5/3))) * (1 + (3/4) * (B0_prima - 4) * (((V0 / V)**(7/3)) - 1)),5)

def c_a(V):
    return c_a_V0 + alpha*(1 - (V / V0)) + beta*((1 - (V / V0))**2)

## Importando dataframe sobre el que se trabajará

df = pd.read_csv('valores_minimos.csv')
df.drop(columns=['Energia_min'],inplace=True)
df = df.append({'Volumen':V0,'c_a_min':c_a(V0)},ignore_index=True)
df.sort_values(by='Volumen',inplace=True)
df['Presion'] = df['Volumen'].apply(P)
df

## Exportamos el dataframe anterior a un csv

df.to_csv('volumen-c_a-presion.csv', index=False)

## Regresando a directorio principal

os.chdir(directorio_principal)
