==> notebooks-scripts/22-creando-y-generando-archivos-band.ipynb <==
# Creando y generando archivos de entrada .in para Quantum Expresso

## Importando liberías a utilizar

import os
import pandas as pd
import numpy as np

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

scf_path = '../data/processed/scf'
os.chdir(scf_path)

# Visualizando datos con los que se trabajan

df = pd.read_csv('../presion-Volumen-c_a-valores-especificos.csv')
df

## Definiendo función para obtener el valor del lattice parametro $a$ a partir de la siguiente ecuación

$$
V = \frac{\sqrt{3}}{2}a^3(c/a)\quad \Rightarrow \quad a= \sqrt[3]{\frac{2V}{\sqrt{3}\cdot(c/a)}}
$$

Considerando que conocemos el valor de $V$ y $c/a$.

def a(V:float,c_a:float) -> float:
    return round(np.cbrt((2 * V) / (np.sqrt(3) * c_a)),5)

# Anexando columna para valores de $a$

df_tolist = df.loc[:,['Volumenes','c_a']].to_numpy().tolist()

a_values = list()
for df_tonumpy_i in df_tolist:
    a_value = a(df_tonumpy_i[0],df_tonumpy_i[1])
    a_values.append(a_value)

df['a'] = a_values
df

# Definiendo función para generar archivos .in

def gen_bands_file(a_value,c_a_value):
   crystal = r'{crystal}'
   crystal_b = r'{crystal_b}'
   bands_file = f''' &CONTROL
    calculation='bands',
    prefix='mob2',
    pseudo_dir='/home/marcocarmonaga/Apps/qe-7.0/pseudo',
    outdir='./tmp',
!    wf_collect=.true.,
 /
 &SYSTEM
    ibrav =  4,
    celldm(1) = {a_value},
    celldm(3) = {c_a_value},
    nat =  3,
    ntyp = 2,
    nbnd = 20,
    ecutwfc = 36.0,
 !   occupations = 'tetrahedra',
 !   occupations = 'tetrahedra_opt',
 !   occupations = 'smearing',
 !   smearing = 'gauss',
 !   degauss = 0.015,
 /
 &ELECTRONS
    diagonalization = 'davidson'
    mixing_mode = 'plain',
    conv_thr =  1.0d-8
    mixing_beta = 0.7 
 /
ATOMIC_SPECIES
   Mo  95.95  Mo_ONCV_PBE-1.2.upf
   B   10.81  B_ONCV_PBE-1.2.upf
ATOMIC_POSITIONS {crystal}
   Mo 0.00 0.00 0.00
   B .6666666666666666  .3333333333333333  0.5
   B .3333333333333333  .6666666666666666  0.5
K_POINTS {crystal_b}
8
   0.0000000000     0.0000000000     0.0000000000     40 !G
   0.0000000000     0.5000000000     0.0000000000     20 !M
   0.3333333333     0.3333333333     0.0000000000     40 !K
   0.0000000000     0.0000000000     0.0000000000     40 !G
   0.0000000000     0.0000000000     0.5000000000     40 !A
   0.0000000000     0.5000000000     0.5000000000     40 !L
   0.3333333333     0.3333333333     0.5000000000     40 !H
   0.0000000000     0.0000000000     0.5000000000     0  !A
'''
   return bands_file

# Anexando columna de texto para archivo .in

c_a_a_values = df.loc[:,['c_a','a']].to_numpy().tolist()

nscf_text_list = list()
for c_a_a_value in c_a_a_values:
    nscf_text = gen_bands_file(c_a_a_value[1],c_a_a_value[0])
    nscf_text_list.append(nscf_text)

df['in_file'] = nscf_text_list
df

# Creando directorios y archivos .in

df_tolist = df.to_numpy().tolist()

df_tolist

for df_tolist_i in df_tolist:
    os.chdir(str(df_tolist_i[0]))
    with open('mob2_bands.in','w') as bands_file:
            bands_file.writelines(df_tolist_i[4])
            bands_file.close()
    os.chdir('../')
