==> notebooks-scripts/30-creando-y-generando-archivos-para-calculo-de-interfase-y-fononico-48Ry.ipynb <==
# Creando y generando archivos de entrada para obtener la densidad de carga del estado base a 48Ry

## Importando liberías a utilizar

import os

## Definiendo directorio principal

directorio_principal = '/home/marcocarmonaga/Documents/MoB_2-Study-Properties_Quantum_expresso/notebooks-scripts'
os.chdir(directorio_principal)

## Cambiando a carpeta sobre la que se trabajará (directorio secundario)

scf_path = '../data/processed/calculo-de-interfase-y-fonones-48Ry'
os.chdir(scf_path)

# Defininiendo valores de ecutwfc y fonones

fonones = ['01', '02', '03', '04', '05', '06']
k_points = ['18','24','30','36']

# Definiendo función para generar archivos .in

def gen_file_interfase(k_point:str):
   crystal = r'{crystal}'
   file_gen = f''' &CONTROL
    calculation='scf',
    prefix='mob2',
    pseudo_dir='/home/marcocarmonaga/Apps/qe-7.0/pseudo',
    outdir='./tmp',
    wf_collect=.true.,
 /
 &SYSTEM
    ibrav =  4,
    celldm(1) = 5.8733350421208156,
    celldm(3) = 1.07051490,
    nat =  3,
    ntyp = 2,
    nbnd = 20,
    ecutwfc = 48.0,
    ecutrho = 672.0,
    occupations = 'smearing',
    smearing = 'gauss',
    degauss = 0.015,
 /
 &ELECTRONS
    diagonalization = 'davidson'
    mixing_mode = 'plain',
    conv_thr =  1.0d-8
    mixing_beta = 0.7 
 /
ATOMIC_SPECIES
   Mo  92.906  Nb1.00Mo0.00.upf
   B   10.81   B_ONCV_PBE-1.2.upf
ATOMIC_POSITIONS {crystal}
   Mo 0.00 0.00 0.00
   B .6666666666666666  .3333333333333333  0.5
   B .3333333333333333  .6666666666666666  0.5
K_POINTS automatic
   {k_point} {k_point} {k_point}   0 0 1'''
   return file_gen

def gen_file_fonones(fonon:str):
   file_gen = f'''Phonon calc for qp{fonon}
 &INPUTPH
    prefix='mob2',
    tr2_ph=1.0d-14,
    outdir='./tmp',
    low_directory_check=.true.,
    amass(1)=92.906,
    reduce_io=.true.,
    ldisp=.true.,
!    lqdir=.true.,
    nq1=3,
    nq2=3,
    nq3=3,
    start_q={fonon}, last_q={fonon},
    fildyn="x000.dyn",
 /'''
   return file_gen

for k_point in k_points:
        name_file = str(k_point).replace('.','_')
        os.mkdir(name_file)
        os.chdir(name_file)
        with open('mob2_interfase.in','w') as file:
                file.writelines(gen_file_interfase(k_point))
        for fonon in fonones:
                with open(f'mob2_fonon_{fonon}.in','w') as file:
                        file.writelines(gen_file_fonones(fonon))
        os.mkdir('tmp')
        os.chdir('../')
